Vitastic/requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Flask==2.1.1
idna==3.3
imageio==2.16.1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
networkx==2.7.1
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.7
python-dotenv==0.20.0
PyWavelets==1.3.0
requests==2.27.1
scikit-image==0.19.2
scipy==1.8.0
soupsieve==2.3.1
tifffile==2022.3.25
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
